requirement.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
